Save the shm to the storage before restarting workers.
