+++ dlrover/python/tests/test_ckpt_saver.py
@@ -12,6 +12,7 @@
 # limitations under the License.
 
 import os
+import tempfile
 import unittest
 
 import numpy as np
@@ -19,11 +20,13 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from dlrover.python.common.multi_process import SharedMemory
+from dlrover.python.common.multi_process import SharedMemory, SharedQueue
 from dlrover.python.elastic_agent.torch.ckpt_saver import (
+    CheckpointSaver,
+    NoShardingCheckpointEngine,
     NoShardingSaver,
-    convert_torch_dtype_to_numpy,
-    traverse_state_dict,
+    _convert_torch_dtype_to_numpy,
+    _traverse_state_dict,
 )
 
 
@@ -52,7 +55,7 @@ class SimpleNet(nn.Module):
 
 class CheckpointSaverTest(unittest.TestCase):
     def test_close_saver(self):
-        saver = NoShardingSaver()
+        saver = NoShardingSaver("test_ckpt")
         saver._tensor_shm = SharedMemory(name="test", create=True, size=1024)
         saver.close()
         saver.close()
@@ -67,15 +70,34 @@ class CheckpointSaverTest(unittest.TestCase):
             model=model.state_dict(),
             step=step,
         )
-        new_dict = traverse_state_dict(state_dict, visitor)
+        new_dict = _traverse_state_dict(state_dict, visitor)
         self.assertEqual(new_dict, state_dict)
 
     def test_convert_torch_dtype_to_numpy(self):
-        np_dtype = convert_torch_dtype_to_numpy(torch.float32)
+        np_dtype = _convert_torch_dtype_to_numpy(torch.float32)
         self.assertEqual(np_dtype, np.float32)
 
-        np_dtype = convert_torch_dtype_to_numpy(torch.float)
+        np_dtype = _convert_torch_dtype_to_numpy(torch.float)
         self.assertEqual(np_dtype, np.float32)
 
-        np_dtype = convert_torch_dtype_to_numpy(torch.int32)
+        np_dtype = _convert_torch_dtype_to_numpy(torch.int32)
         self.assertEqual(np_dtype, np.int32)
+
+    def test_save_to_storage(self):
+        model = SimpleNet()
+        step = 100
+        state_dict = dict(
+            model=model.state_dict(),
+            step=step,
+        )
+        with tempfile.TemporaryDirectory() as tmpdir:
+            CheckpointSaver._saver_instance = NoShardingSaver(tmpdir)
+            sq = SharedQueue(name="factory", create=True)
+            saving_engine = NoShardingCheckpointEngine(tmpdir)
+            saving_engine.save_to_memory(state_dict, step)
+            saver: CheckpointSaver = CheckpointSaver.get_ckpt_saver()
+            saver._tensor_shm = SharedMemory(name=saver._shm_name)
+            saver.save_shm_to_storage()
+            ckpt_files = os.listdir(tmpdir)
+            self.assertEqual(len(ckpt_files), 1)
+            sq.close()
