+++ dlrover/python/tests/test_elastic_training_agent.py
@@ -29,6 +29,10 @@ from dlrover.python.elastic_agent.master_client import (
     build_master_client,
 )
 from dlrover.python.elastic_agent.monitor.training import TorchTrainingMonitor
+from dlrover.python.elastic_agent.torch.ckpt_saver import (
+    CheckpointSaver,
+    NoShardingSaver,
+)
 from dlrover.python.elastic_agent.torch.training import (
     ElasticLaunchConfig,
     ElasticTrainingAgent,
@@ -203,6 +207,7 @@ class ElasticTrainingAgentRunTest(unittest.TestCase):
             master_addr=master_addr,
             local_addr=self.config.local_addr,
         )
+        CheckpointSaver._saver_instance = NoShardingSaver("/tmp/test")
 
     def tearDown(self):
         self._master.stop()
@@ -270,6 +275,7 @@ class ElasticTrainingAgentRunTest(unittest.TestCase):
             start_method=self.config.start_method,
             log_dir=self.config.log_dir,
         )
+        agent._save_ckpt_to_storage()
         agent._stop_workers_to_restart()
